ffi: tidy descrobject.rs
